Remove nested class from file. Refactor convertdocument logic
